# hw12_13_14_15_calendar/api/openapi/schema.yaml
openapi: 3.1.0
info:
  title: Calendar REST API
  version: 0.1.0
paths:
  /events:
    post:
      summary: Create new event
      responses:
        "200":
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEvent'
    put:
      summary: Update event
      responses:
        "200":
          description: Event updated
        "404":
          description: Event not found
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
    get:
      summary: Get list of events
      responses:
        "200":
          description: Retrieves a list of events for specified period
      parameters:
        - name: period
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/EventsPeriod'
        - name: date
          in: query
          required: true
          schema:
            type: integer
            format: int64
            description: Period start (unix timestamp)

  /events/{id}:
    delete:
      summary: Delete an event by ID
      operationId: deleteEventsID
      responses:
        "200":
          description: Event successfully deleted
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: event id


components:
  schemas:
    Event:
      allOf:
        - $ref: '#/components/schemas/NewEvent'
        - type: object
          required:
            - id
          properties:
            id:
              type: string

    NewEvent:
      type: object
      required:
        - title
        - startTime
        - duration
        - description
        - notifyBefore
      properties:
        title:
          type: string
          description: Event title
        startTime:
          type: integer
          format: int64
          description: Event start date (unix timestamp)
        duration:
          type: integer
          description: Duration of an event (in minutes)
        description:
          type: string
          description: Event description
        notifyBefore:
          type: integer
          description: Amount of minutes to notify user before the event


    EventsPeriod: # query parameter
      type: string
      enum: [ "day", "week", "month" ]

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string