# hw12_13_14_15_calendar/pkg/api/openapi/types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package openapi

// Defines values for EventsPeriod.
const (
	Day   EventsPeriod = "day"
	Month EventsPeriod = "month"
	Week  EventsPeriod = "week"
)

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// Event defines model for Event.
type Event struct {
	// Description Event description
	Description string `json:"description"`

	// Duration Duration of an event (in minutes)
	Duration int    `json:"duration"`
	Id       string `json:"id"`

	// NotifyBefore Amount of minutes to notify user before the event
	NotifyBefore int `json:"notifyBefore"`

	// StartTime Event start date (unix timestamp)
	StartTime int64 `json:"startTime"`

	// Title Event title
	Title string `json:"title"`
}

// EventsPeriod defines model for EventsPeriod.
type EventsPeriod string

// NewEvent defines model for NewEvent.
type NewEvent struct {
	// Description Event description
	Description string `json:"description"`

	// Duration Duration of an event (in minutes)
	Duration int `json:"duration"`

	// NotifyBefore Amount of minutes to notify user before the event
	NotifyBefore int `json:"notifyBefore"`

	// StartTime Event start date (unix timestamp)
	StartTime int64 `json:"startTime"`

	// Title Event title
	Title string `json:"title"`
}

// GetEventsParams defines parameters for GetEvents.
type GetEventsParams struct {
	Period EventsPeriod `form:"period" json:"period"`
	Date   int64        `form:"date" json:"date"`
}

// PostEventsJSONRequestBody defines body for PostEvents for application/json ContentType.
type PostEventsJSONRequestBody = NewEvent

// PutEventsJSONRequestBody defines body for PutEvents for application/json ContentType.
type PutEventsJSONRequestBody = Event